api: put holdings_at_dates into new helpers module
